+++ Documents.Database/Entity/DocumentPrediction.cs
@@ -4,9 +4,9 @@ public class DocumentPrediction
 {
     public int Id { get; set; }
 
-    public int DocumentId { get; set; }
+    public long DocumentId { get; set; }
 
     public required string Value { get; set; }
 
-    public double Prob { get; set; }
+    public float Prob { get; set; }
 }
